moved json files to reference data folder
