+++ src/screens/Profile/Information/PaymentInformation.tsx
@@ -1,10 +1,10 @@
 import {InformationTexts, useTranslation} from '@atb/translations';
 import React from 'react';
-import pa from './paymentInformation.json';
+import paymentInfo from '../../../reference-data/paymentInformation.json';
 import Info, {InformationElement} from './Information';
 export default function PaymentInformation() {
   const {t, language} = useTranslation();
-  const data = pa[language];
+  const data = paymentInfo[language];
 
   return (
     <Info
